+++ src/headers/neonio.h
@@ -31,8 +31,8 @@ typedef FILE* NeStream;
 
 NeStream NeStream_open(char* name, char* mode);
 void NeStream_close(NeStream stream);
-void NeStream_write(NeStream stream, void* data, int size);
-bool NeStream_read(NeStream stream, void* data, int size);
+void NeStream_write(NeStream stream, void* data, size_t size);
+bool NeStream_read(NeStream stream, void* data, size_t size);
 
 void removeZeros(char* string);
 bool is_integer(char* string);
